move scripts to different file
